Extract Importer Step 3: reduce coupling
we can reduce the coupling between ProductImporter
and Product by removing the dependency on Product.new
order and number of arguments. Now we can make importers
with different csv structures without changing Product,
and we can add attributes to Product without (necessarily)
changing ProductImporter classes.

We don't need to add any tests to ProductImporter, because we
have not changed the public interface