Extract Importer Step 4: identify public interface
by identify 'create' as a message that ProductImporter needs to
send to a product, we hide the details of product persistence from
ProductImporter. Now if Product changes how it stores a collection
of products, ProductImporter does not need to change

Once again, we don't need to add tests to ProductImporter because
we have not changed its public interface.