Extract Importer Step 5: method level SRP
SRP applies to methods as well. The reason is not just
to make code more readable or easy to understand, but
to enable future change (by inheritance for example).

By extracting the csv delimiter, the code that converts
a csv line to product attributes, the code that casts
values to their correct types, and the code that creates
a product from an attributes hash to their own methods,
we are able to easily create new importers and independently
change how each of those responsibilities are fulfilled.